moved tests to test_manifest file
